Update script2.js
Added a bit of optimization so that the game is checked for a winner only after at least 5 total moves (meaning the minimum of 3 moves has been made by X). Console prints out who is the winner, this needs to be changed so that the winner is displayed to the page.